Implement Admin Service And fixing some issues

+++ src/main/java/com/jets/clerks/rest/ClerkController.java
@@ -2,10 +2,12 @@ package com.jets.clerks.rest;
 
 import com.jets.clerks.dao.ClerkDao;
 import com.jets.clerks.dto.ClerkPostDto;
+import com.jets.login.dao.CheckerDao;
 
 import jakarta.ws.rs.Consumes;
 import jakarta.ws.rs.DELETE;
 import jakarta.ws.rs.GET;
+import jakarta.ws.rs.HeaderParam;
 import jakarta.ws.rs.NotFoundException;
 import jakarta.ws.rs.POST;
 import jakarta.ws.rs.PUT;
@@ -14,24 +16,34 @@ import jakarta.ws.rs.PathParam;
 import jakarta.ws.rs.Produces;
 import jakarta.ws.rs.core.MediaType;
 import jakarta.ws.rs.core.Response;
+
 @Path("clerks")
 public class ClerkController {
     @GET
-    public Response getAllClerks() {
-        ClerkDao clerkDao = new ClerkDao();
-        var clerks = clerkDao.getAllClerks();
-        if (clerks != null) {
-            return Response.ok().entity(clerks).build();
+    public Response getAllClerks(@HeaderParam("LoginId") String uuid) {
+        CheckerDao checkerDao = new CheckerDao();
+        var emp = checkerDao.getLoggedInEmployee(uuid);
+        if (emp.getRole().equalsIgnoreCase("admin")) {
+            ClerkDao clerkDao = new ClerkDao();
+            var clerks = clerkDao.getAllClerks();
+            if (clerks != null) {
+                return Response.ok().entity(clerks).build();
+            }
         }
         throw new NotFoundException("clerk not found");
     }
+
     @GET
     @Path("{id}")
-    public Response getClerkById(@PathParam("id") int id) {
-        ClerkDao clerkDao = new ClerkDao();
-        var clerks = clerkDao.getClerkById(id);
-        if (clerks != null) {
-            return Response.ok().entity(clerks).build();
+    public Response getClerkById(@PathParam("id") int id, @HeaderParam("LoginId") String uuid) {
+        CheckerDao checkerDao = new CheckerDao();
+        var emp = checkerDao.getLoggedInEmployee(uuid);
+        if (emp.getRole().equalsIgnoreCase("admin")) {
+            ClerkDao clerkDao = new ClerkDao();
+            var clerks = clerkDao.getClerkById(id);
+            if (clerks != null) {
+                return Response.ok().entity(clerks).build();
+            }
         }
         throw new NotFoundException("clerk not found");
     }
@@ -39,8 +51,11 @@ public class ClerkController {
     @POST
     @Consumes(MediaType.APPLICATION_JSON)
     @Produces(MediaType.APPLICATION_JSON)
-    public Response addClerk(ClerkPostDto clerkPostDto){
-        if(clerkPostDto.getRole().equalsIgnoreCase("clerk")){
+    public Response addClerk(ClerkPostDto clerkPostDto, @HeaderParam("LoginId") String uuid) {
+        CheckerDao checkerDao = new CheckerDao();
+        var emp = checkerDao.getLoggedInEmployee(uuid);
+        if (emp.getRole().equalsIgnoreCase("admin")) {
+
             ClerkDao clerkDao = new ClerkDao();
             clerkDao.addClerk(clerkPostDto);
             return Response.ok().entity("clerk is added").build();
@@ -52,10 +67,14 @@ public class ClerkController {
     @PUT
     @Path("{id}")
     @Consumes(MediaType.APPLICATION_JSON)
-    public Response updateClerk(@PathParam("id") int id,ClerkPostDto clerkPostDto){
-        if(clerkPostDto.getRole().equalsIgnoreCase("clerk")){
+    public Response updateClerk(@PathParam("id") int id, ClerkPostDto clerkPostDto,
+            @HeaderParam("LoginId") String uuid) {
+        CheckerDao checkerDao = new CheckerDao();
+        var emp = checkerDao.getLoggedInEmployee(uuid);
+        if (emp.getRole().equalsIgnoreCase("admin")) {
+
             ClerkDao clerkDao = new ClerkDao();
-            clerkDao.updateClerk(id,clerkPostDto);
+            clerkDao.updateClerk(id, clerkPostDto);
             return Response.ok().entity("clerk is updated").build();
         }
         throw new NotFoundException("clerk not found");
@@ -63,15 +82,17 @@ public class ClerkController {
 
     @DELETE
     @Path("{id}")
-    public Response deleteClerkById(@PathParam("id") int id){
-        ClerkDao clerkDao = new ClerkDao();
-       Boolean isDeleted= clerkDao.deleteClerkById(id);
-       if(isDeleted){
-        return Response.ok().entity("clerk is deleted").build();
-       }
-       throw new NotFoundException("clerk not found");
+    public Response deleteClerkById(@PathParam("id") int id, @HeaderParam("LoginId") String uuid) {
+        CheckerDao checkerDao = new CheckerDao();
+        var emp = checkerDao.getLoggedInEmployee(uuid);
+        if (emp.getRole().equalsIgnoreCase("admin")) {
+            ClerkDao clerkDao = new ClerkDao();
+            Boolean isDeleted = clerkDao.deleteClerkById(id);
+            if (isDeleted) {
+                return Response.ok().entity("clerk is deleted").build();
+            }
+        }
+        throw new NotFoundException("clerk not found");
     }
-
-
 
 }
